refactor: Refatora o Ninja Controller para incluir o ResponseEntity nas rotas, customizando as respostas das requisições.

+++ src/main/java/com/diego/CadastroDeNinjas/Ninjas/NinjaController.java
@@ -1,5 +1,7 @@
 package com.diego.CadastroDeNinjas.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,28 +15,43 @@ public class NinjaController {
     }
 
     @PostMapping("/add")
-    public NinjaDTO createNinja(@RequestBody NinjaDTO ninjaDTO){
-        return ninjaService.addNinja(ninjaDTO);
+    public ResponseEntity<String> createNinja(@RequestBody NinjaDTO ninjaDTO){
+        NinjaDTO ninja = ninjaService.addNinja(ninjaDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Ninja " + ninja.getName() + " criado com sucesso!");
     }
 
     @GetMapping("/list")
-    public List<NinjaDTO> listAllNinjas(){
-        return ninjaService.listAllNinjas();
+    public ResponseEntity<List<NinjaDTO>> listAllNinjas(){
+        return ResponseEntity.ok().body(ninjaService.listAllNinjas());
     }
 
     @GetMapping("/list/{id}")
-    public NinjaDTO listNinjaId(@PathVariable Long id){
-        return ninjaService.listNinjaById(id);
+    public ResponseEntity<?> listNinjaId(@PathVariable Long id){
+        if(ninjaService.listNinjaById(id) != null){
+            NinjaDTO ninja = ninjaService.listNinjaById(id);
+            return ResponseEntity.ok().body(ninja);
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O ID " + id + " não existe.");
+        }
     }
 
     @PutMapping("/change/{id}")
-    public NinjaDTO changeNinja(@PathVariable Long id, @RequestBody NinjaDTO ninjaDTO){
-        return ninjaService.changeNinja(id, ninjaDTO);
+    public ResponseEntity<String> changeNinja(@PathVariable Long id, @RequestBody NinjaDTO ninjaDTO){
+        if(ninjaService.listNinjaById(id) != null){
+            ninjaService.deleteNinja(id);
+            return ResponseEntity.ok().body("Ninja com o ID " + id + " alterado com sucesso!");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O ID " + id + " não existe.");
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteNinja(@PathVariable Long id){
-        ninjaService.deleteNinja(id);
+    public ResponseEntity<String> deleteNinja(@PathVariable Long id){
+        if(ninjaService.listNinjaById(id) != null){
+            ninjaService.deleteNinja(id);
+            return ResponseEntity.ok().body("Ninja com o ID " + id + " deletado com sucesso!");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O ID " + id + " não existe.");
+        }
     }
-
 }
